+++ packages/@expo/cli/CHANGELOG.md
@@ -17,6 +17,7 @@
 ### 💡 Others
 
 - Update usage of `@expo/server` API ([#39894](https://github.com/expo/expo/pull/39894) by [@kitten](https://github.com/kitten))
+- Rename to `@expo/server` to `expo-server` ([#40087](https://github.com/expo/expo/pull/40087) by [@kitten](https://github.com/kitten))
 
 ### ⚠️ Notices